requirements.txt
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
decorator==4.1.2
gitdb2==2.0.3
GitPython==2.1.7
idna==2.6
mysqlclient==1.3.12
networkx==2.0
numpy==1.13.3
pycparser==2.18
pygit2==0.26.0
PyGithub==1.35
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
regex==2017.9.23
requests==2.18.4
scipy==1.0.0
setproctitle==1.1.10
six==1.11.0
smmap2==2.0.3
SQLAlchemy==1.1.14
SQLAlchemy-Utils==0.32.21
thespian==3.8.0
Unidecode==0.4.21
urllib3==1.22
wrapt==1.10.11
